fix: #65 DesignKit은 third party를 dependency로 들고있지 않아서, Trash관련 Component에서 NMap 관련 오류
